Add Procfile and nltk downloads to run app
For now, the application is just being run using Flask's built-in
server. Probably don't need anything fancier than that for the
purposes of a hackathon project, though we could switch to
gunicorn or something later

On application load, installs individual nltk resources for
tokenizing, part-of-speech tagging, named-entity chunking, and a
corpus of known words. This is done to avoid downloading the complete
set of nltk models and corpora, which would take a while