Add module to predict gender of extracted names
Several changes were required to accomplish this:
* The `genderize` library was added as a dependency, which wraps
the genderize.io API. It is free for up to 1000 requests/day, and
seems to have reasonably good accuracy at predicting gender for
first names
* Because gender prediction is optimized for _first names only_, the
name phrases that are output by the name-extractor (e.g. Senator
Chuck Schumer) need to be gender-predicted in a way that makes sense,
without having any systematic way of determining which part of the
name phrase is the given name. Since genderize.io passes back both a
"count", which I assume means the number of times the name crops up
in their source corpus, we do a prediction on all separate words in the
name phrase and assign the gender that goes with the highest-count result.
    - genderize.io also returns "probability", so we could do our own
probability filtering if we wanted. for now, we're just going with
the default
* In order to handle weird character issues, `unidecode` was added
as a dependency to translate unicode into ASCII before processing