Use local gender-prediction library instead of genderize.io
* The `gender-detector` library was added as a dependency. It is not
a super high-performance gender predictor, but it is FOSS and pip-
installable, and faster than doing lots of API calls, so it'll do
* Because gender prediction is optimized for _first names only_, the
name phrases that are output by the name-extractor (e.g. Senator
Chuck Schumer) need to be gender-predicted in a way that makes sense,
without having any systematic way of determining which part of the
name phrase is the given name. After running through a few actual
examples of news articles from Vox and the New York Times, I noticed
that the `gender-detector` library marks lots of titles and surnames
as "male" (which is itself somewhat of a depressing commentary).
Given that, I decided to return "male" if _all_ of the parts of the
name phrase that gave a valid prediction (i.e. not "unknown"), but to
return "female" if _any_ of the parts of the name phrase return "female".
Artisanal testing suggests that this is a reasonable first-pass strategy